refactor: remove redundant index casts and collection temporary

+++ src/engine/builtins/array.rs
@@ -876,7 +876,7 @@ impl Runtime {
                 )?))
             }
             NativeConversion::Value(InternalDefineResult::RejectedOrdinary(target)) => {
-                let key = self.property_key_for_index(index as u64)?;
+                let key = self.property_key_for_index(index)?;
                 let array_length_read_only =
                     if let ArrayOwnKey::Index(index) = self.array_own_key(&target, &key)? {
                         let (length, writable) = self.array_length_state(&target)?;
@@ -909,7 +909,7 @@ impl Runtime {
         index: u64,
         value: Value,
     ) -> Result<NativeConversion<InternalDefineResult>, RuntimeError> {
-        let key = self.property_key_for_index(index as u64)?;
+        let key = self.property_key_for_index(index)?;
         let descriptor = OrdinaryPropertyDescriptor {
             value: DescriptorField::Present(value),
             writable: DescriptorField::Present(true),
@@ -1152,7 +1152,7 @@ impl Runtime {
                 Completion::Throw(value) => return Ok(Completion::Throw(value)),
             };
         for index in 0..length {
-            let key = self.property_key_for_index(index as u64)?;
+            let key = self.property_key_for_index(index)?;
             let mut value = match self.get_property_in_realm(realm, &source, &key)? {
                 Completion::Return(value) => value,
                 Completion::Throw(value) => return Ok(Completion::Throw(value)),
@@ -1555,7 +1555,7 @@ impl Runtime {
                     )?));
                 }
                 for source_index in 0..length {
-                    let key = self.property_key_for_index(source_index as u64)?;
+                    let key = self.property_key_for_index(source_index)?;
                     let present = match self.has_property_in_realm(realm, &element, &key)? {
                         Completion::Return(Value::Bool(value)) => value,
                         Completion::Return(_) => {
@@ -2680,7 +2680,7 @@ impl Runtime {
         object: &ObjectRef,
         index: u64,
     ) -> Result<NativeConversion<Option<Value>>, RuntimeError> {
-        let key = self.property_key_for_index(index as u64)?;
+        let key = self.property_key_for_index(index)?;
         let present = match self.has_property_in_realm(realm, object, &key)? {
             Completion::Return(Value::Bool(value)) => value,
             Completion::Return(_) => {
@@ -2728,7 +2728,7 @@ impl Runtime {
                 NativeConversion::Value(value) => value,
                 NativeConversion::Throw(value) => return Ok(Some(value)),
             };
-            let to_key = self.property_key_for_index(to as u64)?;
+            let to_key = self.property_key_for_index(to)?;
             if let Some(value) = value {
                 if let Some(value) = self.set_property_or_throw(realm, object, &to_key, value)? {
                     return Ok(Some(value));
@@ -2756,7 +2756,7 @@ impl Runtime {
         object: &ObjectRef,
         index: u64,
     ) -> Result<Option<Value>, RuntimeError> {
-        let key = self.property_key_for_index(index as u64)?;
+        let key = self.property_key_for_index(index)?;
         match self.internal_delete_property(realm, object, &key)? {
             NativeConversion::Value(true) => Ok(None),
             NativeConversion::Value(false) => Ok(Some(self.new_native_error(
@@ -2871,7 +2871,7 @@ impl Runtime {
             let index = from
                 .checked_add(offset)
                 .ok_or(RuntimeError::Invariant("Array push index overflowed"))?;
-            let key = self.property_key_for_index(index as u64)?;
+            let key = self.property_key_for_index(index)?;
             if let Some(value) = self.set_property_or_throw(realm, &object, &key, value)? {
                 return Ok(Completion::Throw(value));
             }
@@ -2924,14 +2924,14 @@ impl Runtime {
 
             match (lower_value, upper_value) {
                 (lower_value, Some(upper_value)) => {
-                    let lower_key = self.property_key_for_index(lower as u64)?;
+                    let lower_key = self.property_key_for_index(lower)?;
                     if let Some(value) =
                         self.set_property_or_throw(realm, &object, &lower_key, upper_value)?
                     {
                         return Ok(Completion::Throw(value));
                     }
                     if let Some(lower_value) = lower_value {
-                        let upper_key = self.property_key_for_index(upper as u64)?;
+                        let upper_key = self.property_key_for_index(upper)?;
                         if let Some(value) =
                             self.set_property_or_throw(realm, &object, &upper_key, lower_value)?
                         {
@@ -2949,7 +2949,7 @@ impl Runtime {
                     {
                         return Ok(Completion::Throw(value));
                     }
-                    let upper_key = self.property_key_for_index(upper as u64)?;
+                    let upper_key = self.property_key_for_index(upper)?;
                     if let Some(value) =
                         self.set_property_or_throw(realm, &object, &upper_key, lower_value)?
                     {
@@ -3237,7 +3237,7 @@ impl Runtime {
                 slot.value = Value::Undefined;
                 continue;
             }
-            let key = self.property_key_for_index(position as u64)?;
+            let key = self.property_key_for_index(position)?;
             let value = std::mem::replace(&mut slot.value, Value::Undefined);
             if let Some(value) = self.set_property_or_throw(realm, object, &key, value)? {
                 return Ok(Some(value));
@@ -3247,7 +3247,7 @@ impl Runtime {
 
         let mut index = defined_count;
         for _ in 0..undefined_count {
-            let key = self.property_key_for_index(index as u64)?;
+            let key = self.property_key_for_index(index)?;
             if let Some(value) =
                 self.set_property_or_throw(realm, object, &key, Value::Undefined)?
             {
@@ -3543,7 +3543,7 @@ impl Runtime {
             let index = start.checked_add(offset).ok_or(RuntimeError::Invariant(
                 "Array.splice item index overflowed",
             ))?;
-            let key = self.property_key_for_index(index as u64)?;
+            let key = self.property_key_for_index(index)?;
             if let Some(value) = self.set_property_or_throw(realm, &object, &key, value)? {
                 return Ok(Completion::Throw(value));
             }
